refactor: split error.rs
