+++ storages/jsonl-storage/src/error.rs
@@ -1,0 +1,40 @@
+use {gluesql_core::result::Error, std::fmt};
+
+pub trait ResultExt<T, E: ToString> {
+    fn map_storage_err(self) -> Result<T, Error>;
+}
+
+impl<T, E: ToString> ResultExt<T, E> for std::result::Result<T, E> {
+    fn map_storage_err(self) -> Result<T, Error> {
+        self.map_err(|e| e.to_string()).map_err(Error::StorageMsg)
+    }
+}
+
+pub trait OptionExt<T> {
+    fn map_storage_err(self, payload: String) -> Result<T, Error>;
+}
+
+impl<T> OptionExt<T> for std::option::Option<T> {
+    fn map_storage_err(self, payload: String) -> Result<T, Error> {
+        self.ok_or_else(|| payload.to_string())
+            .map_err(Error::StorageMsg)
+    }
+}
+
+pub enum JsonlStorageError {
+    FileNotFound,
+    TableDoesNotExist,
+    ColumnDoesNotExist,
+}
+
+impl fmt::Display for JsonlStorageError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let payload = match self {
+            JsonlStorageError::FileNotFound => "file not found".to_owned(),
+            JsonlStorageError::TableDoesNotExist => "table does not exist".to_owned(),
+            JsonlStorageError::ColumnDoesNotExist => "column does not exist".to_owned(),
+        };
+
+        write!(f, "{}", payload)
+    }
+}
